Add Makefile target to fuzz with AFL
Also add a compact test case. As far as I understand, a good test case
should exercise many code paths while being as small as possible.
There's probably still room for improvement.

The variable $AFL_PATH must point to the directory containing the AFL
binaries. It can be set as an environment variable or passed to make on
the command line.