Fixed entity lookup table.
The old one had many errors.
The new one is derived from the list in the npm entities package.
Since the sequences can now be longer (multi-code-point), we
have bumped the length limit from 4 to 8, which also affects
houdini_html_u.c.

An example of the kind of error that was fixed in given
in jgm/commonmark.js#47: `&ngE;` should be rendered as "≧̸" (U+02267
U+00338), but it's actually rendered as "≧" (which is the same as
`&gE;`).