Better handle libdir
On some OSes we want to install the library in a different location
depending on the system architecture.

For example, on a 64 bits Fedora system, the library should be installed
in /usr/lib64, as opposed to the traditional /usr/lib.

This commit allows the builder to specify their own library installation
directory, using the variable that the Fedora build process already
defines.