Correctly handle list marker followed only by spaces.
This change allows us to pass the new test introduced in
75f231503d2b5854f1ff517402d2751811295bf7.

Previously when a list marker was followed only by spaces,
cmark expected the following content to be indented by
the same number of spaces.  But in this case we should
treat the line just like a blank line and set list padding
accordingly.