New parsing strategy for emph/strong.
The idea is that process_emphasis is run twice, first
looking only at delimiters that are unambiguous openers
or closers, and then looking at delimiters that can play
both roles.

This gives better results for things like

    *foo **bar** baz*

It would have to go with a spec change.  Four spec
tests fail.  There is a minor performance penalty (2%
in our benchmark).

Thanks to Fletcher Penney for useful conversation about
the cases.