Changed process_emphasis to get better results in corner cases.
This will need corresponding spec changes.

The change is this:  when considering matches between an interior
delimiter run (one that can open and can close) and another delimiter
run, we require that the (X mod 2) + (Y mod 2) == 1, where X is
the length of the first one and Y is the length of the second one.

This means that in

*a**b*

the second * can't match the first, for example.