Add exposing list markers *, -, +, or numeric to AST and API.
Prior to this commit the node struct for a list contained a member to
store whatever char was parsed from the markdown, but this value
wasn't exposed through the API, nor was it made available in the
output of the AST in any format.

Following this commit a `cmark_marker_type` struct has been added to
capture various forms of list markers: `CMARK_NUMERIC_MARKER` (for
ordered lists) and `CMARK_ASTERISK_MARKER`, `CMARK_HYPHEN_MARKER`, &
`CMARK_PLUS_MARKER` (for unordered lists). The value of which is
stored in the `marker` member of the `cmark_list` node struct.

This enables two primary features:

1. The bullet marker types are now preserved when converting from
markdown to markdown.
2. The XML representation of the AST now has a `marker="hyphen"` (as
well as numeric, asterisk, and plus) as an attribute on the list
nodes. This is similar to the `delim="period"` attribute available on
ordered list nodes.

Tests have been added and are passing. The original test suite still
passes as well. This commit also updates the racket-lang wrapper to
support these new features as well as fixes a minor bug for a missing
`no_delim` symbol because `NO_DELIM` wasn't defined in cmark.h