Prevent out-of-bound memory access. (#124)
* Prevent out-of-bound memory access.

Before the change, the scan_delimiters function could return max_delims+1.
The scan_delimiters function is called by extensions/strikethrough with:

18:    inline_parser, sizeof(buffer)-1, '~'

where buffer is:

12:    char buffer[101];

So scan_delimiters can return delims = 101. A few lines later, it is
used to 0-terminate the buffer with:

23:    buffer[delims] = 0;

so it effectively does buffer[101] = 0, which is an out-of-bound access.

A test was added. That's the test that I used inside Google to reproduce
the bug, thanks to internal infrastructure (sanitizers) to detect bad
memory access. Unfortunately, I couldnt' reproduce the bug with
cmark-gfm build suite. I tried `make asan test` but it gave nothing. I'm
still including the test for reference.

* Limit the recursion in autolink extension.

This is not really a bug, but it's possible to send an input markdown
consisting of lots of @ signs, and the recursion will cause memory
explosion. The limit depends on the running environment, but there is no
reason to accept arbitrarily long sequence of @, so let's just cut off
at 1000.

* Reverted change