build: use target properties for include paths
This configures the target to setup the include paths publicly for the
library targets in the build interface.  This enables uses of the
targets in the build tree without having to specify the include
directories.  This is particularly useful for use in the export targets,
but also simplifies the rules for the API tests.  The install interface
does not need the include directories as `cmark.h` is installed into
`include` which is a default include path.