Generate and install cmake-config file
Add full cmake support. The project can either be used with
`add_subdirectory` or be installed into the system (or some other
directory) and be found with `find_package(cmark)`. In both cases the
cmake target `cmark::cmark` and/or `cmark::cmark_static` is all that
is needed to be linked.

Previously the cmarkConfig.cmake file was generated, but not installed.
As additional bonus of generation by cmake we get a generated
`cmake-config-version.cmake` file for `find_package()` to search for
the same major version.

The generated config file is position independent, allowing the
installed directory to be copied or moved and still work.

The following four files are generated and installed:
- lib/cmake/cmark/cmark-config.cmake
- lib/cmake/cmark/cmark-config-version.cmake
- lib/cmake/cmark/cmark-targets.cmake
- lib/cmake/cmark/cmark-targets-release.cmake