Fix #383.
Previously cmark incorrectly parsed input like

    *****Hello*world****

because once an emphasis opener failed to match a closer
of the same shape, it was taken off the list of potential
openers.  However, if the reason for the match failure is
the "mod 3 rule," then the possibility remains that a different
closer, later in the string, might match an opener that was
rejected.  So, if the reason for the match failure was the
mod 3 rule, we don't adjust the stack bottom.