Fix additional paragraphs in lists
The new variable LIST_LEVEL can help keep track of how deep we
are in a list, and is currently just used to make sure additional
paragraphs in lists are indented properly.

The formatting generated could be cleaned up, as it throws an '.RS'
and '.RE' around each additional paragraph in a list item.