Properly indent block-level contents of nested lists in man.
Closes #258.

This was prompted by #386 by @nicqrocks, but this solution is
more general; it handles not only nested list items but
paragraphs and otehr block-level content under list items.

The change requires addition of a new field block_number_in_list_item
to cmark_renderer, but this is not a public API change because
the details of cmark_renderer are not exposed.