Fix install paths in libcmark.pc
CMAKE_INSTALL_<dir> can be relative or absolute path, so it is wrong to
prefix CMAKE_INSTALL_PREFIX because if CMAKE_INSTALL_<dir> is set to an
absolute path it will result in a malformed path with two absolute paths
joined together:

    $ cmake -DCMAKE_INSTALL_PREFIX=/foo/bar \
        -D CMAKE_INSTALL_BINDIR=/foo/bar/bin \
        -D CMAKE_INSTALL_LIBDIR=/foo/bar/lib \
        -D CMAKE_INSTALL_INCLUDEDIR=/foo/bar/include \
        ..

    $ cat src/libcmark.pc
    prefix=/foo/bar
    exec_prefix=/foo/bar
    libdir=/foo/bar//foo/bar/lib
    includedir=/foo/bar//foo/bar/include
    ...

The libdir and includedir should be set by either calculating the
directories relative to the prefix or by using the
CMAKE_INSTALL_FULL_<dir> variables from GNUInstallDirs.

The CMAKE_INSTALL_FULL_<dir> variables seem the simpler solution.