use an in-tree header for symbol export information
Relying on CMake's GenerateExportHeader produces a file that is longer
than the one being added here, which for all that is just mostly
defining macros not used here. And after all that, it only contains
macros specific to a single compiler, while failing to consistently
handle GNUC (that always supports symbol visibility even for static
libraries).

Replace this with a more targeted header that is easy to read or include
into external build systems, and which is also more robust than the one
that only exists inside CMake.