Fix quadratic complexity bug.
Previously in parsing the repeated pattern ![[]() the parser
repeatedly scanned to the beginning of the increasingly large
stack of bracketed delimiters, trying to set the link delimiters
to "inactive" to prevent links inside links.

This commit removes the `active` flag from the bracketed delimiters, and
adds a new boolean flag `no_link_openers` on subject. This can be
set to true after a link is matched, telling the parser not to
bother forming a link if it matches an open bracket. It is set
to false when a new link open bracket is added to the stack of
openers. This new approach avoids the need to traverse the stack
of open brackets.