Defer resolution of link reference definitions
We must not remove link reference definitions until we check for list
tightness.  This commit defers resolving of link reference definitions
until finalization of the document.  We still need to eagerly remove
link reference definitions in setext headings to determine whether it is
a setext heading or a thematic break.  So this commit provides slightly
different functions for resolving link reference definitions and
checking if a line is blank for `cmark_strbuf` and `cmark_chunk`.