remove some compiler compatibility checks
`stdbool.h` is part of C99 however was not provided by Visual Studio
2013 until RTM [1]. Remove the check for the header and inline the
include at the usage sites rather than relying on `config.h`.  VS2013
was EOL'ed Apr 9, 2019, with extended support ending Apr 9, 2024.

`HAVE___ATTRIBUTE__` was unused in the codebase and served no purpose.

Remove shims for `snprintf` and `vsnprintf` which were unavailable prior
to VS2015.  As VS2013 is no longer serviced, this reduces complexity in
the project.

[1] https://devblogs.microsoft.com/cppblog/c99-library-support-in-visual-studio-2013/