build: minor tweaks for speeding up the configure phase
Only the C language is used unless we are building the ASAN or UBSAN
build variants. Only check for the C language support, enabling C++
support in the case that the build type is one of those two.

Restructure the file layout to place the custom modules into the
`cmake/modules` directory that is the common layout for CMake based
projects.

While a generally unimportant optimization, it makes it much more clear
what is being checked and why and does reduce the configure time by ~10%
(.25s) on Windows.